requirements.txt
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pillow==11.3.0
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
trimesh==4.7.1
typing_extensions==4.14.1
